requirements.txt
beautifulsoup4==4.11.1
build==0.7.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
et-xmlfile==1.1.0
Flask==2.1.1
google-api-core==2.7.2
google-auth==2.6.6
google-cloud==0.34.0
google-cloud-core==2.3.0
google-cloud-pubsub==2.12.0
googleapis-common-protos==1.56.0
grpc-google-iam-v1==0.12.4
grpcio==1.44.0
grpcio-status==1.44.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
lxml==4.8.0
MarkupSafe==2.1.1
numpy==1.22.3
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
pep517==0.12.0
proto-plus==1.20.3
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
soupsieve==2.3.2.post1
tomli==2.0.1
urllib3==1.26.9
Werkzeug==2.1.1
xlrd==2.0.1
zipp==3.8.0
